yn_cm

+++ erpandcomm/src/main/java/kr/spring/stock/vo/CurrentStockVO.java
@@ -16,6 +16,6 @@ public class CurrentStockVO {
     private Date last_in_date;       // 마지막 입고일
     private Date last_out_date;      // 마지막 출고일
     private long last_movement_num;  // 마지막 재고이동번호
-    private long version_num;        // 동시성 제어용 버전
+    private int version_num;        // 동시성 제어용 버전
     private Date updated_date;       // 마지막 거래/수정일
 }
